change return from int to double (#2)

+++ src/main/java/com/example/springdocker/MyMathCalc.java
@@ -17,7 +17,8 @@ public class MyMathCalc {
         return a*b;
     }
 
-    public float divide (int a, int b){
+    public double divide (int a, int b){
+
         return a/b;
     }
 }
